Move Teams Use Cases Into Namespace
