+++ spec/use_cases/teams/fetch_spec.rb
@@ -1,4 +1,4 @@
-describe UseCases::FetchTeams do
+describe UseCases::Teams::Fetch do
   it 'Calls execute on the gateway' do
     teams_gateway = double(execute: [])
     expect(teams_gateway).to receive(:execute).once
